Complete macOS export cancellation flow
