+++ apps/macos/Sources/OpenRecorderMac/InspectorViews.swift
@@ -1,0 +1,386 @@
+import AVFoundation
+import AppKit
+import CoreGraphics
+import SwiftUI
+import UniformTypeIdentifiers
+
+struct SettingsInspector: View {
+    @EnvironmentObject private var model: AppModel
+    @Binding var borderRadius: Double
+    @Binding var padding: Double
+    @Binding var shadow: Double
+    @Binding var backgroundBlur: Double
+    @Binding var loopCursor: Bool
+    @Binding var cursorSize: Double
+    @Binding var cursorSmoothing: Double
+    var recordingSession: RecordingSession?
+
+    @State private var activeTab: InspectorTab = .appearance
+
+    var body: some View {
+        HStack(spacing: 0) {
+            inspectorRail
+            inspectorContent
+        }
+        .background(Color.studioPanel.opacity(0.86), in: RoundedRectangle(cornerRadius: 10))
+        .overlay {
+            RoundedRectangle(cornerRadius: 10)
+                .stroke(Color.studioBorder)
+        }
+        .shadow(color: Color.black.opacity(0.18), radius: 18, y: 12)
+    }
+
+    private func openExternal(_ urlString: String) {
+        guard let url = URL(string: urlString) else { return }
+        NSWorkspace.shared.open(url)
+    }
+
+    private var inspectorRail: some View {
+        VStack(spacing: 8) {
+            ForEach(InspectorTab.allCases) { tab in
+                InspectorRailButton(tab: tab, isActive: activeTab == tab) {
+                    activeTab = tab
+                }
+            }
+        }
+        .frame(width: 56)
+        .frame(maxHeight: .infinity, alignment: .top)
+        .padding(.vertical, 10)
+        .background(Color.white.opacity(0.025))
+        .overlay(alignment: .trailing) {
+            Rectangle()
+                .fill(Color.studioBorder)
+                .frame(width: 1)
+        }
+    }
+
+    private var inspectorContent: some View {
+        VStack(spacing: 0) {
+            ScrollView {
+                VStack(alignment: .leading, spacing: 12) {
+                    inspectorHeader
+                    tabContent
+                }
+                .padding(12)
+            }
+
+            Rectangle()
+                .fill(Color.studioBorder)
+                .frame(height: 1)
+
+            inspectorFooter
+        }
+    }
+
+    private var inspectorHeader: some View {
+        HStack(spacing: 9) {
+            Image(systemName: activeTab.symbolName)
+                .font(.system(size: 14, weight: .semibold))
+                .foregroundStyle(Color.brand)
+                .frame(width: 30, height: 30)
+                .background(Color.brand.opacity(0.12), in: RoundedRectangle(cornerRadius: 7))
+
+            VStack(alignment: .leading, spacing: 3) {
+                Text(activeTab.title)
+                    .font(.system(size: 14, weight: .semibold))
+                Text(activeTab.subtitle)
+                    .font(.system(size: 10))
+                    .foregroundStyle(.secondary)
+                    .fixedSize(horizontal: false, vertical: true)
+            }
+
+            Spacer(minLength: 0)
+
+            Text(activeTab.id)
+                .font(.system(size: 9, weight: .semibold))
+                .foregroundStyle(.secondary)
+                .padding(.horizontal, 6)
+                .padding(.vertical, 3)
+                .background(Color.white.opacity(0.055), in: RoundedRectangle(cornerRadius: 5))
+        }
+        .padding(10)
+        .background(Color.white.opacity(0.045), in: RoundedRectangle(cornerRadius: 8))
+        .overlay {
+            RoundedRectangle(cornerRadius: 8)
+                .stroke(Color.white.opacity(0.06))
+        }
+    }
+
+    private var inspectorFooter: some View {
+        HStack(spacing: 8) {
+            InspectorFooterButton(title: "Report Bug", symbolName: "ladybug") {
+                openExternal("https://github.com/imbhargav5/open-recorder/issues/new/choose")
+            }
+            InspectorFooterButton(title: "Star on GitHub", symbolName: "star") {
+                openExternal("https://github.com/imbhargav5/open-recorder")
+            }
+        }
+        .padding(12)
+        .background(Color.white.opacity(0.025))
+    }
+
+    @ViewBuilder
+    private var tabContent: some View {
+        switch activeTab {
+        case .appearance:
+            InspectorSlider(title: "Shadow", valueText: "\(Int(shadow * 100))%", value: $shadow, range: 0...1, step: 0.01)
+            InspectorSlider(title: "Roundness", valueText: "\(Int(borderRadius))px", value: $borderRadius, range: 0...25, step: 0.5)
+            InspectorSlider(title: "Padding", valueText: "\(Int(padding))%", value: $padding, range: 0...100, step: 1)
+            InspectorSlider(title: "Background Blur", valueText: String(format: "%.1fpx", backgroundBlur), value: $backgroundBlur, range: 0...8, step: 0.25)
+            StudioButton(hitTarget: .rounded(8), action: {}) {
+                Label("Crop Video", systemImage: "crop")
+                    .font(.system(size: 11, weight: .semibold))
+                    .frame(maxWidth: .infinity)
+                    .frame(height: 34)
+                    .background(Color.white.opacity(0.055), in: RoundedRectangle(cornerRadius: 8))
+            }
+            BackgroundPalette()
+        case .cursor:
+            InspectorSwitch(title: "Show Cursor", isOn: $model.showCursor)
+            InspectorSwitch(title: "Loop Cursor", isOn: $loopCursor)
+            InspectorSlider(title: "Size", valueText: String(format: "%.2fx", cursorSize), value: $cursorSize, range: 0.5...10, step: 0.05)
+            InspectorSlider(title: "Smoothing", valueText: String(format: "%.2f", cursorSmoothing), value: $cursorSmoothing, range: 0...2, step: 0.01)
+        case .camera:
+            InspectorSwitch(title: "Facecam", isOn: .constant(recordingSession?.facecamVideoPath != nil), isInteractive: false)
+            InspectorSlider(title: "Facecam Size", valueText: "24%", value: .constant(24), range: 12...40, step: 1)
+            InspectorSlider(title: "Border Width", valueText: "4px", value: .constant(4), range: 0...16, step: 1)
+            if let path = recordingSession?.facecamVideoPath {
+                SessionAssetRow(title: "Facecam File", path: path)
+            }
+            PositionGrid()
+        case .audio:
+            InspectorSwitch(title: "Mute Preview", isOn: .constant(false), isInteractive: false)
+            InspectorSlider(title: "Volume", valueText: "100%", value: .constant(1), range: 0...1, step: 0.01)
+            if let sourceName = recordingSession?.sourceName {
+                SessionAssetRow(title: "Source", path: sourceName)
+            }
+        }
+    }
+}
+
+struct SessionAssetRow: View {
+    var title: String
+    var path: String
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 4) {
+            Text(title)
+                .font(.system(size: 10, weight: .semibold))
+                .foregroundStyle(.secondary)
+            Text(path)
+                .font(.system(size: 10, weight: .medium, design: .monospaced))
+                .lineLimit(2)
+                .textSelection(.enabled)
+        }
+        .padding(10)
+        .frame(maxWidth: .infinity, alignment: .leading)
+        .background(Color.white.opacity(0.035), in: RoundedRectangle(cornerRadius: 8))
+    }
+}
+
+struct InspectorRailButton: View {
+    var tab: InspectorTab
+    var isActive: Bool
+    var action: () -> Void
+
+    var body: some View {
+        StudioButton(hitTarget: .rounded(9), help: tab.title, action: action) {
+            Image(systemName: tab.symbolName)
+                .font(.system(size: 15, weight: .semibold))
+                .frame(width: 40, height: 40)
+                .foregroundStyle(isActive ? Color.brand : Color.secondary)
+                .background(isActive ? Color.brand.opacity(0.15) : Color.clear, in: RoundedRectangle(cornerRadius: 9))
+                .overlay {
+                    RoundedRectangle(cornerRadius: 9)
+                        .stroke(isActive ? Color.brand.opacity(0.24) : Color.clear, lineWidth: 1)
+                }
+        }
+    }
+}
+
+struct InspectorFooterButton: View {
+    var title: String
+    var symbolName: String
+    var action: () -> Void
+
+    var body: some View {
+        StudioButton(hitTarget: .rounded(7), action: action) {
+            Label(title, systemImage: symbolName)
+                .font(.system(size: 10, weight: .medium))
+                .frame(maxWidth: .infinity)
+                .frame(height: 30)
+                .foregroundStyle(.secondary)
+                .background(Color.white.opacity(0.035), in: RoundedRectangle(cornerRadius: 7))
+        }
+    }
+}
+
+enum InspectorTab: CaseIterable, Identifiable {
+    case appearance
+    case cursor
+    case camera
+    case audio
+
+    var id: String { title }
+
+    var title: String {
+        switch self {
+        case .appearance: "Appearance"
+        case .cursor: "Cursor"
+        case .camera: "Camera"
+        case .audio: "Audio"
+        }
+    }
+
+    var subtitle: String {
+        switch self {
+        case .appearance: "Frame styling, background, crop, and composition."
+        case .cursor: "Cursor visibility and motion effects."
+        case .camera: "Facecam overlay settings."
+        case .audio: "Master preview and MP4 export audio."
+        }
+    }
+
+    var symbolName: String {
+        switch self {
+        case .appearance: "slider.horizontal.3"
+        case .cursor: "cursorarrow"
+        case .camera: "camera"
+        case .audio: "speaker.wave.2"
+        }
+    }
+}
+
+struct InspectorGroup<Content: View>: View {
+    var title: String
+    var symbolName: String
+    @ViewBuilder var content: Content
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 10) {
+            Label(title, systemImage: symbolName)
+                .font(.system(size: 10, weight: .semibold))
+                .foregroundStyle(.secondary)
+                .textCase(.uppercase)
+
+            VStack(alignment: .leading, spacing: 10) {
+                content
+            }
+        }
+        .padding(10)
+        .background(Color.white.opacity(0.035), in: RoundedRectangle(cornerRadius: 8))
+        .overlay {
+            RoundedRectangle(cornerRadius: 8)
+                .stroke(Color.white.opacity(0.05))
+        }
+    }
+}
+
+struct InspectorSlider: View {
+    var title: String
+    var valueText: String
+    @Binding var value: Double
+    var range: ClosedRange<Double>
+    var step: Double
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 7) {
+            HStack {
+                Text(title)
+                    .font(.system(size: 10, weight: .medium))
+                    .foregroundStyle(.secondary)
+                Spacer()
+                Text(valueText)
+                    .font(.system(size: 10, weight: .medium, design: .monospaced))
+                    .foregroundStyle(Color.secondary.opacity(0.78))
+            }
+            ElasticSlider(value: $value, range: range, step: step)
+                .accessibilityLabel(title)
+        }
+        .padding(10)
+        .background(Color.white.opacity(0.045), in: RoundedRectangle(cornerRadius: 8))
+        .overlay {
+            RoundedRectangle(cornerRadius: 8)
+                .stroke(Color.white.opacity(0.05))
+        }
+    }
+}
+
+struct InspectorSwitch: View {
+    var title: String
+    @Binding var isOn: Bool
+    var isInteractive = true
+
+    var body: some View {
+        HStack {
+            Text(title)
+                .font(.system(size: 11, weight: .medium))
+                .foregroundStyle(.secondary)
+            Spacer()
+            Toggle("", isOn: $isOn)
+                .labelsHidden()
+                .toggleStyle(.switch)
+                .allowsHitTesting(!isInteractive)
+        }
+        .padding(10)
+        .background(Color.white.opacity(0.045), in: RoundedRectangle(cornerRadius: 8))
+        .overlay {
+            RoundedRectangle(cornerRadius: 8)
+                .stroke(Color.white.opacity(0.05))
+        }
+        .rectangularHitTarget()
+        .onTapGesture {
+            guard isInteractive else { return }
+            isOn.toggle()
+        }
+    }
+}
+
+struct BackgroundPalette: View {
+    private let colors: [Color] = [.red, .yellow, .green, .white, .blue, .orange, .purple, .pink, .cyan, .black]
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 8) {
+            Label("Background", systemImage: "paintpalette")
+                .font(.system(size: 11, weight: .semibold))
+                .foregroundStyle(.secondary)
+            LazyVGrid(columns: Array(repeating: GridItem(.flexible(), spacing: 6), count: 5), spacing: 6) {
+                ForEach(colors.indices, id: \.self) { index in
+                    RoundedRectangle(cornerRadius: 6)
+                        .fill(colors[index])
+                        .frame(height: 28)
+                        .overlay {
+                            RoundedRectangle(cornerRadius: 6)
+                                .stroke(Color.white.opacity(0.12))
+                        }
+                }
+            }
+        }
+        .padding(10)
+        .background(Color.white.opacity(0.045), in: RoundedRectangle(cornerRadius: 8))
+        .overlay {
+            RoundedRectangle(cornerRadius: 8)
+                .stroke(Color.white.opacity(0.05))
+        }
+    }
+}
+
+struct PositionGrid: View {
+    var body: some View {
+        VStack(alignment: .leading, spacing: 8) {
+            Text("Position")
+                .font(.system(size: 10, weight: .medium))
+                .foregroundStyle(.secondary)
+            LazyVGrid(columns: Array(repeating: GridItem(.flexible(), spacing: 5), count: 3), spacing: 5) {
+                ForEach(0..<9, id: \.self) { index in
+                    RoundedRectangle(cornerRadius: 5)
+                        .fill(index == 8 ? Color.brand.opacity(0.28) : Color.white.opacity(0.06))
+                        .frame(height: 28)
+                }
+            }
+        }
+        .padding(10)
+        .background(Color.white.opacity(0.045), in: RoundedRectangle(cornerRadius: 8))
+    }
+}
+
